Extract common validation logic to an abstract class and implement ServletContextListener
Fix for #3303
